+++ Source/Krypton Components/Krypton.Toolkit/Values/InputPulsingBorderColorValues.cs
@@ -1,4 +1,4 @@
-#region BSD License
+﻿#region BSD License
 /*
  *
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
@@ -12,6 +12,12 @@ namespace Krypton.Toolkit;
 /// <summary>
 /// Storage for input control pulsing border colors.
 /// </summary>
+/// <remarks>
+/// Control-owned instances inherit unset colors from
+/// <see cref="KryptonManager.PulsingBorderValues"/>.Colors. Assigning a color stores a local
+/// override; <c>Reset*</c> (and <see cref="Reset"/>) clears the override so the global value is
+/// used again. The manager instance does not inherit.
+/// </remarks>
 [TypeConverter(typeof(InputPulsingBorderColorValuesConverter))]
 public class InputPulsingBorderColorValues : Storage
 {
@@ -25,23 +31,42 @@ public class InputPulsingBorderColorValues : Storage
 
     #region Instance Fields
 
-    private Color _color1 = DefaultColor1;
-    private Color _color2 = DefaultColor2;
-    private Color _highlightColor = DefaultHighlightColor;
+    private readonly bool _inheritFromGlobal;
+    private Color? _color1;
+    private Color? _color2;
+    private Color? _highlightColor;
 
     #endregion
 
     #region Identity
 
     /// <summary>
-    /// Initialize a new instance of the InputGlowingBorderColorValues class.
+    /// Initialize a new instance of the <see cref="InputPulsingBorderColorValues"/> class.
     /// </summary>
     /// <param name="needPaint">Delegate for notifying paint requests.</param>
-    public InputPulsingBorderColorValues(NeedPaintHandler? needPaint) 
+    public InputPulsingBorderColorValues(NeedPaintHandler? needPaint)
+        : this(needPaint, inheritFromGlobal: true)
+    {
+    }
+
+    /// <summary>
+    /// Initialize a new instance of the <see cref="InputPulsingBorderColorValues"/> class.
+    /// </summary>
+    /// <param name="needPaint">Delegate for notifying paint requests.</param>
+    /// <param name="inheritFromGlobal">
+    /// When <see langword="true"/>, unset colors read from
+    /// <see cref="KryptonManager.PulsingBorderValues"/>.Colors. Pass <see langword="false"/> for
+    /// the manager's own global instance.
+    /// </param>
+    internal InputPulsingBorderColorValues(NeedPaintHandler? needPaint, bool inheritFromGlobal)
     {
         NeedPaint = needPaint;
+        _inheritFromGlobal = inheritFromGlobal;
 
-        Reset();
+        if (!inheritFromGlobal)
+        {
+            Reset();
+        }
     }
 
     /// <inheritdoc />
@@ -68,24 +93,17 @@ public class InputPulsingBorderColorValues : Storage
     [TypeConverter(typeof(ColorConverter))]
     public Color Color1
     {
-        get => _color1;
+        get => GetInherited(_color1, colors => colors.Color1, DefaultColor1);
 
-        set
-        {
-            if (_color1 != value)
-            {
-                _color1 = value;
-                PerformNeedPaint(true);
-            }
-        }
+        set => SetLocal(ref _color1, value);
     }
 
-    private bool ShouldSerializeColor1() => _color1 != DefaultColor1;
+    private bool ShouldSerializeColor1() => ShouldSerializeLocal(_color1, Color1, DefaultColor1);
 
     /// <summary>
     /// Resets the Color1 property to its default value.
     /// </summary>
-    public void ResetColor1() => Color1 = DefaultColor1;
+    public void ResetColor1() => ResetLocal(ref _color1, DefaultColor1);
 
     #endregion
 
@@ -99,24 +117,17 @@ public class InputPulsingBorderColorValues : Storage
     [TypeConverter(typeof(ColorConverter))]
     public Color Color2
     {
-        get => _color2;
+        get => GetInherited(_color2, colors => colors.Color2, DefaultColor2);
 
-        set
-        {
-            if (_color2 != value)
-            {
-                _color2 = value;
-                PerformNeedPaint(true);
-            }
-        }
+        set => SetLocal(ref _color2, value);
     }
 
-    private bool ShouldSerializeColor2() => _color2 != DefaultColor2;
+    private bool ShouldSerializeColor2() => ShouldSerializeLocal(_color2, Color2, DefaultColor2);
 
     /// <summary>
     /// Resets the Color2 property to its default value.
     /// </summary>
-    public void ResetColor2() => Color2 = DefaultColor2;
+    public void ResetColor2() => ResetLocal(ref _color2, DefaultColor2);
 
     #endregion
 
@@ -130,34 +141,80 @@ public class InputPulsingBorderColorValues : Storage
     [TypeConverter(typeof(ColorConverter))]
     public Color HighlightColor
     {
-        get => _highlightColor;
+        get => GetInherited(_highlightColor, colors => colors.HighlightColor, DefaultHighlightColor);
 
-        set
-        {
-            if (_highlightColor != value)
-            {
-                _highlightColor = value;
-                PerformNeedPaint(true);
-            }
-        }
+        set => SetLocal(ref _highlightColor, value);
     }
 
-    private bool ShouldSerializeHighlightColor() => _highlightColor != DefaultHighlightColor;
+    private bool ShouldSerializeHighlightColor() =>
+        ShouldSerializeLocal(_highlightColor, HighlightColor, DefaultHighlightColor);
 
     /// <summary>
     /// Resets the HighlightColor property to its default value.
     /// </summary>
-    public void ResetHighlightColor() => HighlightColor = DefaultHighlightColor;
+    public void ResetHighlightColor() => ResetLocal(ref _highlightColor, DefaultHighlightColor);
 
     #endregion
 
     #region Reset
 
+    /// <summary>
+    /// Resets all colors. Control instances clear local overrides and inherit from
+    /// <see cref="KryptonManager.PulsingBorderValues"/> again; the global instance restores factory defaults.
+    /// </summary>
     public void Reset()
     {
         ResetColor1();
         ResetColor2();
         ResetHighlightColor();
+    }
+
+    #endregion
+
+    #region Implementation
+
+    private InputPulsingBorderColorValues? InheritSource =>
+        _inheritFromGlobal && !ReferenceEquals(this, KryptonManager.PulsingBorderValues.Colors)
+            ? KryptonManager.PulsingBorderValues.Colors
+            : null;
+
+    private Color GetInherited(Color? local, Func<InputPulsingBorderColorValues, Color> read, Color factory)
+    {
+        if (local.HasValue)
+        {
+            return local.Value;
+        }
+
+        InputPulsingBorderColorValues? source = InheritSource;
+        return source != null ? read(source) : factory;
+    }
+
+    private bool ShouldSerializeLocal(Color? local, Color effective, Color factory) =>
+        _inheritFromGlobal ? local.HasValue : effective != factory;
+
+    private void SetLocal(ref Color? field, Color value)
+    {
+        if (!field.HasValue || field.Value != value)
+        {
+            field = value;
+            PerformNeedPaint(true);
+        }
+    }
+
+    private void ResetLocal(ref Color? field, Color factory)
+    {
+        if (_inheritFromGlobal)
+        {
+            if (field.HasValue)
+            {
+                field = null;
+                PerformNeedPaint(true);
+            }
+
+            return;
+        }
+
+        SetLocal(ref field, factory);
     }
 
     #endregion